[IDP-462] refactor not_started tests hierarchy
